refactor: renamed the file to index in fullView
