+++ frontend/src/container/GridGraphLayout/Graph/index.tsx
@@ -45,7 +45,7 @@ import { v4 } from 'uuid';
 import { LayoutProps } from '..';
 import EmptyWidget from '../EmptyWidget';
 import WidgetHeader from '../WidgetHeader';
-import FullView from './FullView/index.metricsBuilder';
+import FullView from './FullView';
 import { FullViewContainer, Modal } from './styles';
 
 function GridCardGraph({
